refactor the code based on comment
