+++ sql/core/src/test/scala/org/apache/spark/sql/execution/streaming/TextSocketStreamSuite.scala
@@ -20,12 +20,18 @@ package org.apache.spark.sql.execution.streaming
 import java.io.{IOException, OutputStreamWriter}
 import java.net.ServerSocket
 import java.sql.Timestamp
+import java.util.Optional
 import java.util.concurrent.LinkedBlockingQueue
+
+import scala.collection.JavaConverters._
+import scala.collection.mutable.ListBuffer
 
 import org.scalatest.BeforeAndAfterEach
 
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.AnalysisException
+import org.apache.spark.sql.{AnalysisException, Row}
+import org.apache.spark.sql.sources.v2.DataSourceV2Options
+import org.apache.spark.sql.sources.v2.streaming.reader.MicroBatchReader
 import org.apache.spark.sql.streaming.StreamTest
 import org.apache.spark.sql.test.SharedSQLContext
 import org.apache.spark.sql.types.{StringType, StructField, StructType, TimestampType}
@@ -44,10 +50,15 @@ class TextSocketStreamSuite extends StreamTest with SharedSQLContext with Before
       source.stop()
       source = null
     }
+    if (batchReader != null) {
+      batchReader.stop()
+      batchReader = null
+    }
   }
 
   private var serverThread: ServerThread = null
   private var source: Source = null
+  private var batchReader: MicroBatchReader = null
 
   test("basic usage") {
     serverThread = new ServerThread()
@@ -85,6 +96,65 @@ class TextSocketStreamSuite extends StreamTest with SharedSQLContext with Before
       // Try stopping the source to make sure this does not block forever.
       source.stop()
       source = null
+    }
+  }
+
+  test("V2 basic usage") {
+    serverThread = new ServerThread()
+    serverThread.start()
+
+    val provider = new TextSocketSourceProvider
+    val options = new DataSourceV2Options(
+      Map("host" -> "localhost", "port" -> serverThread.port.toString).asJava)
+    batchReader = provider.createMicroBatchReader(Optional.empty(), "", options)
+
+    val schema = batchReader.readSchema()
+    assert(schema === StructType(StructField("value", StringType) :: Nil))
+
+    failAfter(streamingTimeout) {
+      serverThread.enqueue("hello")
+      batchReader.setOffsetRange(Optional.empty(), Optional.empty())
+      while (batchReader.getEndOffset.asInstanceOf[TextSocketOffset].offset == -1L) {
+        batchReader.setOffsetRange(Optional.empty(), Optional.empty())
+        Thread.sleep(10)
+      }
+      withSQLConf("spark.sql.streaming.unsupportedOperationCheck" -> "false") {
+        val offset1 = batchReader.getEndOffset
+        val batch1 = new ListBuffer[Row]
+        batchReader.createReadTasks().asScala.map(_.createDataReader()).foreach { r =>
+          while (r.next()) {
+            batch1.append(r.get())
+          }
+        }
+        assert(batch1.map(_.getAs[String](0)) === Seq("hello"))
+
+        serverThread.enqueue("world")
+        while (batchReader.getEndOffset === offset1) {
+          batchReader.setOffsetRange(Optional.of(offset1), Optional.empty())
+          Thread.sleep(10)
+        }
+        val offset2 = batchReader.getEndOffset
+        val batch2 = new ListBuffer[Row]
+        batchReader.createReadTasks().asScala.map(_.createDataReader()).foreach { r =>
+          while (r.next()) {
+            batch2.append(r.get())
+          }
+        }
+        assert(batch2.map(_.getAs[String](0)) === Seq("world"))
+
+        batchReader.setOffsetRange(Optional.empty(), Optional.of(offset2))
+        val both = new ListBuffer[Row]
+        batchReader.createReadTasks().asScala.map(_.createDataReader()).foreach { r =>
+          while (r.next()) {
+            both.append(r.get())
+          }
+        }
+        assert(both.map(_.getAs[String](0)) === Seq("hello", "world"))
+      }
+
+      // Try stopping the source to make sure this does not block forever.
+      batchReader.stop()
+      batchReader = null
     }
   }
 
@@ -131,24 +201,94 @@ class TextSocketStreamSuite extends StreamTest with SharedSQLContext with Before
     }
   }
 
+  test("V2 timestamped usage") {
+    serverThread = new ServerThread()
+    serverThread.start()
+
+    val provider = new TextSocketSourceProvider
+    val options = new DataSourceV2Options(Map("host" -> "localhost",
+      "port" -> serverThread.port.toString, "includeTimestamp" -> "true").asJava)
+    batchReader = provider.createMicroBatchReader(Optional.empty(), "", options)
+
+    val schema = batchReader.readSchema()
+    assert(schema === StructType(StructField("value", StringType) ::
+      StructField("timestamp", TimestampType) :: Nil))
+
+    failAfter(streamingTimeout) {
+      serverThread.enqueue("hello")
+      batchReader.setOffsetRange(Optional.empty(), Optional.empty())
+      while (batchReader.getEndOffset.asInstanceOf[TextSocketOffset].offset == -1L) {
+        batchReader.setOffsetRange(Optional.empty(), Optional.empty())
+        Thread.sleep(10)
+      }
+      withSQLConf("spark.sql.streaming.unsupportedOperationCheck" -> "false") {
+        val offset1 = batchReader.getEndOffset
+        val batch1 = new ListBuffer[Row]
+        batchReader.createReadTasks().asScala.map(_.createDataReader()).foreach { r =>
+          while (r.next()) {
+            batch1.append(r.get())
+          }
+        }
+        assert(batch1.map(_.getAs[String](0)) === Seq("hello"))
+        val batch1Stamp = batch1.map(_.getAs[Timestamp](1)).head
+
+        serverThread.enqueue("world")
+        while (batchReader.getEndOffset === offset1) {
+          batchReader.setOffsetRange(Optional.of(offset1), Optional.empty())
+          Thread.sleep(10)
+        }
+        val offset2 = batchReader.getEndOffset
+        val batch2 = new ListBuffer[Row]
+        batchReader.createReadTasks().asScala.map(_.createDataReader()).foreach { r =>
+          while (r.next()) {
+            batch2.append(r.get())
+          }
+        }
+        assert(batch2.map(_.getAs[String](0)) === Seq("world"))
+        val batch2Stamp = batch2.map(_.getAs[Timestamp](1)).head
+        assert(!batch2Stamp.before(batch1Stamp))
+      }
+
+      // Try stopping the source to make sure this does not block forever.
+      batchReader.stop()
+      batchReader = null
+    }
+  }
+
   test("params not given") {
     val provider = new TextSocketSourceProvider
     intercept[AnalysisException] {
       provider.sourceSchema(sqlContext, None, "", Map())
     }
     intercept[AnalysisException] {
+      provider.createMicroBatchReader(Optional.empty(), "",
+        new DataSourceV2Options(Map.empty[String, String].asJava))
+    }
+    intercept[AnalysisException] {
       provider.sourceSchema(sqlContext, None, "", Map("host" -> "localhost"))
     }
     intercept[AnalysisException] {
+      provider.createMicroBatchReader(Optional.empty(), "",
+        new DataSourceV2Options(Map("host" -> "localhost").asJava))
+    }
+    intercept[AnalysisException] {
       provider.sourceSchema(sqlContext, None, "", Map("port" -> "1234"))
+    }
+    intercept[AnalysisException] {
+      provider.createMicroBatchReader(Optional.empty(), "",
+        new DataSourceV2Options(Map("port" -> "1234").asJava))
     }
   }
 
   test("non-boolean includeTimestamp") {
     val provider = new TextSocketSourceProvider
+    val params = Map("host" -> "localhost", "port" -> "1234", "includeTimestamp" -> "fasle")
     intercept[AnalysisException] {
-      provider.sourceSchema(sqlContext, None, "", Map("host" -> "localhost",
-      "port" -> "1234", "includeTimestamp" -> "fasle"))
+      provider.sourceSchema(sqlContext, None, "", params)
+    }
+    intercept[AnalysisException] {
+      val a = new DataSourceV2Options(params.asJava)
+      provider.createMicroBatchReader(Optional.empty(), "", a)
     }
   }
 
@@ -157,13 +297,21 @@ class TextSocketStreamSuite extends StreamTest with SharedSQLContext with Before
     val userSpecifiedSchema = StructType(
       StructField("name", StringType) ::
       StructField("area", StringType) :: Nil)
+    val params = Map("host" -> "localhost", "port" -> "1234")
     val exception = intercept[AnalysisException] {
       provider.sourceSchema(
         sqlContext, Some(userSpecifiedSchema),
         "",
-        Map("host" -> "localhost", "port" -> "1234"))
+        params)
     }
     assert(exception.getMessage.contains(
+      "socket source does not support a user-specified schema"))
+
+    val exception1 = intercept[AnalysisException] {
+      provider.createMicroBatchReader(
+        Optional.of(userSpecifiedSchema), "", new DataSourceV2Options(params.asJava))
+    }
+    assert(exception1.getMessage.contains(
       "socket source does not support a user-specified schema"))
   }
 
@@ -172,6 +320,10 @@ class TextSocketStreamSuite extends StreamTest with SharedSQLContext with Before
     val parameters = Map("host" -> "localhost", "port" -> "0")
     intercept[IOException] {
       source = provider.createSource(sqlContext, "", None, "", parameters)
+    }
+    intercept[IOException] {
+      batchReader = provider.createMicroBatchReader(
+        Optional.empty(), "", new DataSourceV2Options(parameters.asJava))
     }
   }
